Module reikna::aliquot

Module for working with aliquot and divisor sums.

This module contains functions for calculating the aliquot and divisor sums of numbers, along with functions for testing for perfect numbers and similar concepts.

Functions

abundant_number
Return true if n is an abundant number, that is, a number whose aliquot sum is greater than itself.

aliquot_sum
Return the aliquot sum of a positive integer n, that is, the sum of all of n's proper divisors.

amicable_number
Return true if n is a member of an amicable pair, that is, a pair of numbers whose aliquot sums equal each other.

deficient_number
Return true if n is a deficient number, that is, a number whose aliquot sum is less than itself.

divisor_sum
Return the divisor sum of a positive integer n, that is, the sum of all of n's divisors.

perfect_number
Return true if n is a perfect number, that is, a number whose aliquot sum is equal to itself.

quasiperfect_number
Return true if n is a quasiperfect number, that is, a number whose aliquot sum is exactly one greater than itself.

sociable_number
Return true if n is a sociable number, that is, a number whose aliquot sums form a cyclic pattern, e.g.

superperfect_number
Return true if n is a superperfect number, that is, a number which satisfies