Function reikna::aliquot::quasiperfect_number

pub fn quasiperfect_number(n: u64) -> bool

Return true if n is a quasiperfect number, that is, a number whose aliquot sum is exactly one greater than itself.

No quasiperfect numbers are known.

Panics

Panics if n is zero.

Examples

use reikna::aliquot::quasiperfect_number;
assert_eq!(quasiperfect_number(28), false);
assert_eq!(quasiperfect_number(145_688), false);