Function reikna::continued_fraction::expand_fraction_ntimes

pub fn expand_fraction_ntimes(fraction: &ContinuedFraction, n: u64) -> (u64, u64)

Expand the continued fraction fraction n times, storing the result as a fraction in a double tuple of u64.

The result tuple is formatted as:

(numerator, denominator)

Finite fractions should be expanded with n = 1, infinite fractions should be expanded with an n large enough to gain the desired precision.

Note that is n is large or the continued fraction is very long, the u64s representing the numerator and denominator may overflow.

Panics

Panics if n is zero or if fraction is empty.

Examples

use reikna::continued_fraction::expand_fraction_ntimes;
assert_eq!(expand_fraction_ntimes(&vec![1, 2], 3), (41, 29));
assert_eq!(expand_fraction_ntimes(&vec![14], 2), (14, 1));