Function reikna::partition::part_wc

pub fn part_wc(n: i64, cache: &mut Vec<u64>) -> u64

Calculate the partition function of n using cache to cache previously calculated values.

This function will recursively calculate the partition function using Euler's generating function and the Pentagonal Number Theorem. This function uses a Vec<u64> to cache previously calculated values, which is necessary to allow the function to run in a reasonable time.

The cache is also useful when calculating multiple partition values, as many of the values only need to be calculated once and can be reused for all other partition calculations.

The cache MUST have a length of at least n + 1.

An empty cache should contain only zeros, otherwise the function will incorrectly assume that the non-zero values hold a cached value, and will use this incorrect value in calculations.

Panics

Panics if n is greater than MAX_PART. This limit is put in place to prevent stack and integer overflows.

Panics if cache.len() is less than n + 1.

Examples

use reikna::partition::part_wc;
let mut cache: Vec<u64> = vec![0; 101];
assert_eq!(part_wc(100, &mut cache), 190569292);
assert_eq!(part_wc(5, &mut cache), 7);
assert_eq!(part_wc(3, &mut cache), 3);