Function reikna::prime_count::prime_count_all

pub fn prime_count_all(data: &Vec<u64>) -> Vec<u64>

Calculate the value of the prime-counting function for each value in data, and return a new Vec<u64> of the results.

The resulting vector has the same size as the input vector.

This function works in fundamentally the same way as prime_count(), with the modification that caches are preserved between calculations. This allows for much faster computation of multiple values.

Panics

Panics if prime_sieve() panics, see the documentation of prime_sieve() for more information.

Examples

use reikna::prime_count::prime_count_all;
assert_eq!(prime_count_all(&vec![1, 2, 3]), vec![0, 1, 2]);
assert_eq!(prime_count_all(&vec![100]), vec![25]);