Module reikna::totient

Module for working with Euler's totient function.

This module has functions for computing the value of the totient function, both for single and multiple values.

Constants

PHI_SYMBOL
Constant string of the uppercase Phi symbol, often used to represent the totient function.

Functions

totient
Calculate the value of Euler's totient function for n.

totient_all
Calculate the value of Euler's totient function for each value in data, and return a new Vec<u64> of the results.